+++ tools/symbol-extractor/run_all_symbols_extractor_tests.js
@@ -23,21 +23,25 @@ const USER_COMMAND = argv._[0];
 // The shell command to query for all tests.
 // Bazel targets for testing goldens
 process.stdout.write('Gathering all symbol extractor targets...');
+
+const BAZEL_QUERY =
+  `'kind(js_test, ${TEST_TARGETS_LOCATION}) ` +
+  `intersect attr("tags", "symbol_extractor", ${TEST_TARGETS_LOCATION})'`;
 const ALL_TEST_TARGETS = spawnSync(
   'yarn',
-  [
-    '-s',
-    'bazel',
-    'query',
-    '--output',
-    'label',
-    `'kind(nodejs_test, ${TEST_TARGETS_LOCATION}) intersect attr("tags", "symbol_extractor", ${TEST_TARGETS_LOCATION})'`,
-  ],
+  ['-s', 'bazel', 'query', '--output', 'label', BAZEL_QUERY],
   {encoding: 'utf8', shell: true, cwd: path.resolve(__dirname, '../..')},
 )
   .stdout.trim()
   .split('\n')
-  .map((line) => line.trim());
+  .map((line) => line.trim())
+  .filter((line) => line.length > 0);
+
+if (ALL_TEST_TARGETS.length === 0) {
+  console.error(`Could not find any symbol test targets matching query: ${BAZEL_QUERY}`);
+  process.exit(1);
+}
+
 process.stdout.clearLine();
 process.stdout.cursorTo(0);
 // Bazel targets for generating goldens
