+++ tools/symbol-extractor/symbol_extractor.mts
